New Camera Implementation
Changed the way the camera functions.  Game camera is now handled by a separate function that will pass x and y to camera() in the game_update() function.

Along with this, I added more map to level 1 in order to test the full bounds of the map and to make sure the camera's right most position is correct.