Item Pickups
Items can now be picked up. Nico checks against the object table first to see if an object can be picked up.  If the object has a touch() function, then we check to see if Nico and that object have their hitboxes colliding. If they do, we run the object's touch function.